# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Jobs",
        "description": "Laravel API documentation for the 'My Jobs' application.",
        "contact": {
            "email": "erick1souza1ago04@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/profiles": {
            "get": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Retrieve a list of profiles",
                "description": "Returns a list of all profiles.",
                "operationId": "profiles_index",
                "responses": {
                    "200": {
                        "description": "A list of profiles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "Profile ID",
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "name": {
                                                        "description": "Profile name",
                                                        "type": "string",
                                                        "example": "Moder"
                                                    },
                                                    "guard_name": {
                                                        "description": "Guard name",
                                                        "type": "string",
                                                        "example": "api"
                                                    },
                                                    "created_at": {
                                                        "description": "Creation timestamp",
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-07-02T23:12:14.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "description": "Update timestamp",
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-07-02T23:12:14.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Create a new profile",
                "description": "Creates a new profile with specified name and permissions.",
                "operationId": "profiles_store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "permissions"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Name of the new profile",
                                        "type": "string",
                                        "example": "Editor"
                                    },
                                    "permissions": {
                                        "type": "array",
                                        "items": {
                                            "description": "Permission name",
                                            "type": "string",
                                            "enum": [
                                                "api.vacancies.index",
                                                "api.vacancies.store",
                                                "api.vacancies.show",
                                                "api.vacancies.update",
                                                "api.vacancies.destroy",
                                                "api.infos.index",
                                                "api.infos.store",
                                                "api.infos.show",
                                                "api.infos.update",
                                                "api.infos.destroy",
                                                "api.auth.login",
                                                "api.auth.create",
                                                "api.auth.logout",
                                                "api.auth.refresh",
                                                "api.auth.me",
                                                "api.auth.infos",
                                                "api.vacancies_user.my_published_vacancies",
                                                "api.vacancies_user.my_applications_vacancies",
                                                "api.vacancies_user.to_apply_or_unapply",
                                                "api.vacancies_user.vacancy_applications"
                                            ]
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The created profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "Profile ID",
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "name": {
                                                        "description": "Profile name",
                                                        "type": "string",
                                                        "example": "Moder"
                                                    },
                                                    "guard_name": {
                                                        "description": "Guard name",
                                                        "type": "string",
                                                        "example": "api"
                                                    },
                                                    "created_at": {
                                                        "description": "Creation timestamp",
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-07-02T23:12:14.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "description": "Update timestamp",
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-07-02T23:12:14.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/profiles/assign_role_for_user/{user_id}": {
            "get": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Assign one role for user",
                "description": "Assign one role for user.",
                "operationId": "assign_role_for_user",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "role_id",
                        "in": "query",
                        "description": "Role to assign.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile assign for user successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Restore user token.",
                "description": "Restore user token.",
                "operationId": "refresh",
                "responses": {
                    "200": {
                        "description": "User information successfully retrieved.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "token"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": 3600
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}